Created Generator interface, separated operation types and tables and C generation code into new files
